refactor: group and rename files in common folder
