+++ src/app.ts
@@ -1,5 +1,5 @@
 import express, { Express } from 'express';
-import { ExceptionFilter } from './common/middlewares/exception-filter';
+import { ErrorFilterMiddleware } from './common/middleware/error-filter-middleware';
 import { injectable, inject } from 'inversify';
 import 'reflect-metadata';
 import { CONTAINER_TYPES } from './common/constants';
@@ -15,7 +15,8 @@ export class App {
 		@inject(CONTAINER_TYPES.APPOINTMENTS_ROUTES) private readonly appointmentsRoutes: IRoutes,
 		@inject(CONTAINER_TYPES.AUTH_ROUTES) private readonly authRoutes: IRoutes,
 		@inject(CONTAINER_TYPES.USER_ROUTES) private readonly userRoutes: IRoutes,
-		@inject(CONTAINER_TYPES.EXCEPTION_FILTER) private readonly exceptionFilter: ExceptionFilter,
+		@inject(CONTAINER_TYPES.ERROR_FILTER_MIDDLEWARE)
+		private readonly exceptionFilter: ErrorFilterMiddleware,
 	) {
 		this.app = express();
