+++ src/common/constants.ts
@@ -25,7 +25,7 @@ export const CONTAINER_TYPES = {
 	USER_REPOSITORY: Symbol.for('USER_REPOSITORY'),
 
 	APP: Symbol.for('APP'),
-	EXCEPTION_FILTER: Symbol.for('EXCEPTION_FILTER'),
+	ERROR_FILTER_MIDDLEWARE: Symbol.for('EXCEPTION_FILTER'),
 
 	APPOINTMENTS_DATA_PROVIDER: Symbol.for('APPOINTMENTS_DATA_PROVIDER'),
 	DB_CONNECTION: Symbol.for('DB_CONNECTION'),
