+++ src/patient/patient-routes.ts
@@ -3,7 +3,7 @@ import { IRoutes } from '../common/types';
 import { injectable, inject } from 'inversify';
 import { CONTAINER_TYPES } from '../common/constants';
 import { IHttpController } from '../common/types';
-import { QueryMapperMiddleware } from '../common/middlewares/query-mapper-middleware';
+import { QueryMapperMiddleware } from '../common/middleware/query-mapper-middleware';
 import { AuthMiddleware } from '../auth/auth-middleware';
 import { iocContainer } from '../inversify.config';
